+++ src/lib/game/levels.ts
@@ -13,6 +13,7 @@ import level11Data from './levels/level-11.json';
 import level12Data from './levels/level-12.json';
 import level13Data from './levels/level-13.json';
 import level14Data from './levels/level-14.json';
+import levelKeys1Data from './levels/level-keys.json';
 
 export const LEVEL_1 = LevelDefinitionSchema.parse(level1Data);
 export const LEVEL_2 = LevelDefinitionSchema.parse(level2Data);
@@ -28,6 +29,7 @@ export const LEVEL_11 = LevelDefinitionSchema.parse(level11Data);
 export const LEVEL_12 = LevelDefinitionSchema.parse(level12Data);
 export const LEVEL_13 = LevelDefinitionSchema.parse(level13Data);
 export const LEVEL_14 = LevelDefinitionSchema.parse(level14Data);
+export const LEVEL_KEYS_1 = LevelDefinitionSchema.parse(levelKeys1Data);
 
 export const LEVELS = [
 	LEVEL_1,
@@ -43,5 +45,6 @@ export const LEVELS = [
 	LEVEL_11,
 	LEVEL_12,
 	LEVEL_13,
-	LEVEL_14
+	LEVEL_14,
+	LEVEL_KEYS_1
 ];